Bug 11833462 - lock_table() shouldn't acquire the lock mutex to check whether a transaction holds a lock on a table
lock_table() needs to check whether a transaction has a (strong enough) lock
on a table, and if it doesn't it needs to create and acquire the requested
lock. The table locks are in the dict_table_t::locks for every transaction
on the system. This can make the check expensive for >= 256 transactions and
increase the pressure on the lock mutex.

With this patch we cache the table locks in trx_t::table_locks. Because the
table locks are only added by the running transaction and removed at
trx_commit() by the owning transaction, it is safe to read the locks
vector without holding any mutex in lock_table().

rb://605 Approved by: Jimmy Yang.