SPJ: Refactoring of how NdbQueryOperationDef objects was released:
Previously NdbQueryOperationDef's was released by either 
::~NdbQueryBuilderImpl() or ::~NdbQueryDefImpl() doing
a 'delete' on the NdbQueryOperationDef being the root of the
query tree and each NdbQueryOperationDef recursively deleting
its children. This had the implicit assumption that the
query tree was in a consistent state - Which it might not be 
in the case where a NdbQueryBuilder factory method failed 
before a query tree had been completed.

It simplify the logic to remove the recursive deletion of children,
and instead iterate the m_operations[] vector to find the 
NdbQueryOperationDef to delete.