Bug#11822517: Refactor call interface of choose_plan()
Part 1 - create class Optimize_table_order.

This patch creates a class that is used to optimize the join order of
tables in a query. The main purpose of the refactoring is to simplify
function interfaces and to encapsulate information needed to perform
optimization in a dedicated class.

The following existing functions are added to the class:
  choose_table_order() - renamed from choose_plan()
  check_interleaving_with_nj()
  advance_sj_state()
  backout_nj_sj_state()
  optimize_straight_join()
  greedy_search()
  best_extension_by_limited_search()
  optimize_wo_join_buffering()
  determine_search_depth()

sql/sql_select.cc
  New class definition, see above for details.

  make_join_statistics() is now taking full responsibility for error
  checking, so there is longer need for testing is_fatal_error() after
  calling this function.

sql/sql_select.h
  Added all_table_map (map of tables in query) to JOIN class.
  cur_embedding_map has been moved to class Optimize_table_order.
  Commented const_table_map, found_const_table_map and outer_join.