Bug#12370950 TABLE_IO_WAITS_SUMMARY_BY_INDEX_USAGE AGGREGATES NON-INSERT DML WRONGLY
Before this fix, table io for UPDATE and DELETE of a row was systematically counted,
in table performance_schema.table_io_waits_summary_by_index_usage,
as an operation using no index.

With this fix, when an index was used to execute an UPDATE or DELETE statement,
the table io for the UPDATE and DELETE are counted against the index used.