Bug#11765860: main.kill fails under valgrind with 'invalid file descriptor' warning
A pre-requisite patch for removing signal-based alarms, aiming
to replace signal-based alarms with standard synchronization
functions in order to simplify the process of killing a slave
thread.

A new pair of mutex and condition variable is introduced
to synchronize thread sleep and finalization. A new mutex is
required because the slave threads are terminated while holding
the slave thread locks (run_lock), which can not be relinquished
during termination as this would affect the lock order.