Bug#11757855 - 49967: built-in libedit doesn't read                       .editrc on linux.
MySQL client when build with libedit support ignores
.editrc at startup.

The reason for this regression was the incluison of a
safety check, issetugid(), which is not available on
some linux platforms.

Fixed by adding an equivalent check for platforms which
have get[e][u|g]id() set of functions.


cmd-line-utils/libedit/el.c:
  Bug#11757855 - 49967: built-in libedit doesn't read
                        .editrc on linux.
  
  Added function calls to check user/group IDs on linux
  systems which does not have issetugid() function.
configure.in:
  Bug#11757855 - 49967: built-in libedit doesn't read
                        .editrc on linux.
  
  Added check for getuid, geteuid, getgid, getegid
  functions.