BUG#11766327: Range analysis should not be done many times               for the same index
(Former bug 59415)
      
The optimizer often choses to access a table through an index
that does not provide the correct ordering for free. To remedy
this, the function test_if_skip_sort_order() is called to see
if another index is as good as the chosen index and at the
same time is able to provide ordering.
            
This implies that test_if_skip_sort_ordering() goes through a
full range analysis (if range access is applicable) to check
whether or not another range access plan should be used
instead of the currently chosen ref/range access method.
        
The problem is that if range analysis is performed and it is
decided that it is not better than whatever we had, the range
analysis will most likely be performed again and again with
the same outcome because test_if_skip_sort_order() is called
from multiple locations.
            
This patch avoids the unnecessarily repeated range analysis
described above by introducing key_map
JOIN_TAB::quick_order_tested which is checked to see
if range analysis has already been performed for a given key.