SPJ: Fixed a regression introduced with recent refactoring of ha_ndbcluster_push.cc:
As we now allow dependencies between (grand-)parents to be added such that all
the dependent parents become (artificial) dependent on each other, we may also
implicit introduce dependencies on outer joined tables.

This is incorrect as it is only inner joined tables which can 
sequentially linked wo/ changing the sematics in the SQL statement.