Avoid mtr warning on Windows during startup
The reason for mtr warning is that collect_mysqld_features() starts mysqld with --datadir=/tmp and this 
directory does not exist on Windows.

Fix : instead of passing --datadir=$opt_vardir/tmp in collect_mysqld_features() just use --datadir=.
mysqld does not need a correct directory, just an existing one, as it is started with --help ---verbose -skip-grant-tables.