Bug#11764086 Bug#56881: Null left operand to NOT IN in WHERE clause behaves differently than real NULL
The query that triggers this problem is an outer join query with
a NOT IN subquery predicate in the WHERE clause.
One of the tables in the outer join contains only one row, and because
it is a MyISAM table, const table optimization is applied to it.

Now the problem: not_null_tables() for the NOT IN predicate reports
that we can ignore NULL-complemented rows for table child, hence the
query is converted from a left join to an inner join. This leads
the query executor to omit the row with id 2 from the evaluation.

As for why the not_null_tables() result for NOT IN is wrong:
NOT IN may return TRUE for a row, even if the left-hand expression
to NOT IN is NULL (as in NULL NOT IN (<empty subquery>)).
This is what happens when the query executor evaluates the row with
id 2 from the parent table: There is no corresponding row in the
child table, a NULL-complemented row should be added, and the
predicate NULL NOT IN (<empty subquery> is evaluated (and it should
return TRUE).

The solution to the problem is to adjust the value of
not_null_tables_cache for Item_in_optimizer::fix_fields().
If the Item_in_subselect member is "top level", meaning that the
original query is an IN predicate, use the accumulated not-null
information for the object, otherwise (the predicate is NOT IN)
remove the set of tables referred from the left-hand expression
from the accumulated not-null information.

mysql-test/include/subquery.inc
  Added test case for bug#11764086.

mysql-test/r/subquery_all.result
mysql-test/r/subquery_all_jcl6.result
mysql-test/r/subquery_nomat_nosj.result
mysql-test/r/subquery_nomat_nosj_jcl6.result
mysql-test/r/subquery_none.result
mysql-test/r/subquery_none_jcl6.result
  Added test results for bug#11764086.

sql/item_cmpfunc.cc
  Adjusted implementation of not_null_tables_cache in
  Item_in_optimizer::fix_fields().

sql/item_func.h
  Added some comments for relevant Item members.