Small refactoring of NdbQueryBuilder's construction of pushed operations (SPJ API) Partly a followup to: mysql.com/commits/135407
To simplify error handling code when building NdbQueryOperationDef's, the allocated
operation is added to NdbQueryBuilders'Vector<NdbQueryOperationDefImpl*> m_operations'
earlier. In case of a failure during query build, the operation will then be 
released together with all other resources allocated so far through this NdbQueryBuilder.

This removed the need for manually delete the allocated operation in case of failures.