BUG#12358775: RPL.RPL_CRASH_SAFE_MASTER FAILS OCCASIONALLY ON PB2
In some of the tests included in the test file, the master is
forced to crash and restart.  In some of these, a slave is trying
to reconnect while the master is down. It can happen that the
master takes more than 10s to restart, thus the slave IO thread
will stop with an error in such cases. Eventually the whole test
fails with a timeout on a sync_slave_with_master operation (which
ends up in timeout in master_wait_pos).

To fix this we stop the slave IO thread before crashing and
restarting the master. We restart the slave IO thread afterward
and then proceed with the synchronization of the slave and check
the data. This does not harm the original assertions, still it
protects the test against slave reconnection timeout.