Refactoring / cleanup of SPJ API after http://lists.mysql.com/commits/136639 (Which allows pushed scan childs to be 'bushy')
The NdbQueryBuilder used to keep track of whether a pushed operation 
already had a scan descendant - Mainly in order to be able to throw
an error of a bushy scan join was constructed.

The only usage of ::hasScanDescendant() besides this was within
NdbQueryImpl::sendFetchMore(). It used it for calculating
which ResultStreams to ::reset() as they would receive new rows.
It turned out that this code was overly complicated, and could be 
simplified - which removed the need for the 'scan descendant' knowledge.

::markScanAncestors(), ::hasScanDescendant() has then become 
obsolete and is therefore removed by this patch.