Fix for Bug#12321461 CRASH IN DSMRR_IMPL::DSMRR_INIT ON SELECT STRAIGHT_JOIN
This crash is caused by a too strict assert that was added in the
patch for Bug#58463. The purpose of this assert was to detect
situations where the optimizer pushes an index condition (ICP) on the
primary index and then later during execution "changes its mind" and
uses a different index. This assert failed to take into account that
the handler might be scanned multiple times during execution of query
and that if MRR is used only on the first initialization of the MRR
scan it will be initialized to use an index. On the second
initialization of the MRR scan it will not have an active index open on
the handler. 

Detailed explanation of how this situation occurs:

The optimizer has decided to scan the second table as a range scan and
pushed an index condition to the handler on the primary key. The range
scan will be done using an DS-MRR scan to read the second table
multiple times (due to join buffering is disabled). This scan is
initialized from QUICK_RANGE_SELECT::reset().

The first time QUICK_RANGE_SELECT::reset() is called it will
initialize the handler to use the primary index. Then it will call
handler::multi_range_read_init() which will set up DS-MRR for this
scan. This is done in DsMrr_impl::dsmrr_init(). The problematic assert
will be evaluated and at this point it passes (since the pushed index
condition is on the same active index as QUICK_RANGE_SELECT::reset()
initialized). The main thing the initialization of the DS-MRR scan
does is to create a "clone" of the handler object that will be used
for some part of the DS-MRR scan. In this cloning process it will
"copy" the pushed index condition from the main handler object to the
clone handler object and set up the clone to do the index scan on the
same index as the main handler was set up to use. Then it will "close"
the active index on the main handler object since this will be used
for reading the records from the base table. 

Then a second scan of the table needs to be done and this causes a
second call to QUICK_RANGE_SELECT::reset(). QUICK_RANGE_SELECT::reset()
checks if the handler has been initialized, and since it has it does
not initialize the index. Then it calls handler::multi_range_read_init()
which calls DsMrr_impl::dsmrr_init(). In DsMrr_impl::dsmrr_init() the
problematic assert will be evaluated. And this time it fails due to the 
handler has no active index (since this was "transferred" to the clone
handler object during the first initialization). Thus the assert fails
even though we have a valid state of the handler object.

The fix for this problem is to make the assert less strict by adding
the case where the handler does not have an active index as a legal
state.

This patch also adds some asserts that validate that the main handler
object and the clone handler object are in a consistent state with
regards to information about pushed index conditions.