Bug 12408132 -  ASSERTION FAILURE IN FILE SYNC0ARR.C LINE 369
Set srv_sync_array_size default size to 1. This is an advanced option and helps
only when number of concurrent threads is greater than equal to 512.

The proper fix is to check all sync array instances for a free slot. This
changeset should fix the immediate problem that was encountered during stress
testing with a small buffer pool and lots of concurrent threads.

Approved by Marko over IRC.