BUG#12416768: RPL_ROW_IGNORABLE_EVENT FAILS SPORADICALLY ON PB2,               RESULT MISMATCH MYSQLBINLOG
There could be a race between the insert delayed thread and the
user thread issuing the DDL statement.

We fix this by making sure in the test case that the DDL statement
is issued only after the delayed thread terminates. We force it
to terminate by resorting to the FLUSH TABLES command. In addition
we also deploy two assertions:

  - assert that there are 0 delayed threads after FLUSH TABLES
  - assert that before the DDL, there was actually something written
    to the binary log as part of the INSERT DELAYED operation