Bug 11766123 - 59163: SINGLE USER SESSION PERFORMANCE IN INNODB
Force a sleep of 1 seconds during flushes even if the server is idle
and we see pending read IOs. What this means is that for workloads
where there is a burst of write activity followed by read only
activity (for example a bulk insert) the dirtied pages will get to
the disk at a slower pace. On the other hand if the server is truly
idle then the flushing will happen in a tight loop. 

Approved by: Marko
rb://562