BUG#12557307 - RPL.RPL_ROTATE_LOGS FAILS IN VALGRIND
The test case removes the master.info repository while it is being used. 
So when the replication is restarted, the missing file is detected and
recreated. This means that the reference to the repository is closed and
the reopened.

However, the IO_CACHE is not closed and is created twice thus leading to
memory leaks. To fix the problem, we always close the IO_CACHE before the
repository is closed.