Fixed recover routine in the context of WL#4832.
The previous routine was considering a transaction as valid after
reading the next begin event or reaching the end of the log. However,
if the next begin was invalid the previous and valid transaction was
removed from the binary log.

We fixed the code by considering a transaction as valid when its
commit is read and is valid.