BUG#12370830: RPL_CHECKSUM FAILED ON PB2 (SERVER CRASH)
The server crashes while trying to printout trace information
when compiled in debug mode. The problem seems to be the same
as the one reported in BUG#11765758, except that it happens in
another test case. 

Fixed the test case by deploying the necessary synchronization 
measures for slave threads before setting the debug variable.