WL5353 Reflecting GCI
Add NDB$ROW_GCI64 and NDB$ROW_AUTHOR optional pseudo columns.

These pseudo columns are only present if requested at table creation time.
Both are variable width - 1 to 31 bits.

Pseudo column NDB$ROW_GCI64 gives access to a higher resolution row GCI/epoch
value.  The number of bits used controls the resolution of the low-word of
the GCI.  Where the bits are insufficient for an actual epoch, reading the 
column returns a saturated value in the low-word, and the normal ROW_GCI value
in the high-word.

Pseudo column NDB$ROW_AUTHOR is automatically set to 0 on INSERT, UPDATE, WRITE
operations.  It can be explicitly set to some other value.  The intention
is to allow the identity of the last user/application to write to a row to be
determined.

ndb_desc is enhanced to indicate the number of bits used for each optional
pseudo column in a table.

ndb_select_all is enhanced to enable the value of these columns to be
output.