WL5353 Primary Cluster Conflict Resolution
Implementation of refreshTuple() mechanism.

A new operation type, refreshTuple() is defined.
When executed, it causes an NdbApi data change event to be 
generated for the affected row.
If the row does not exist, a DELETE event, with the primary key,
is generated.
If the row does exist, an INSERT event, with the primary key and
all values is generated.

A refreshTuple() operation must be the last operation on a particular 
tuple in a transaction.  An error will be returned if any other operation
(including refreshTuple) is defined on a refreshed row in a transaction.

refreshTuple() does not currently support tables with BLOB columns.