Bug 12399395 - ASSERTION FAILURE !MUTEX_OWN(MUTEX) IN SYNC0SYNC.IC LINE 207 4 of 103
When we select a victim during deadlock checking we acquire the trx_t::mutex
on behalf of that transaction. We check for this when discarding the victim
transaction's locks. If it is a deadlock victim there is no need to re-acquire
the trx_t::mutex.

rb://670 Approved by Jimmy Yang.