BUG# 12621935: ASSERTION IN BUF_LRU_FREE_BLOCK WITH UNIV_SYNC_DEBUG
rb://676
Approved by: Jimmy

In buf_LRU_free_block() if we have to free a freshly allocated
buf_page_t, which was allocated for buf_relocate() and is being
freed because the buf_block_t in question got buffer fixed or io fixed,
we'll call buf_buddy_free with hash_lock and block::mutex held. This
breaks latching order and causes an assertion when we try to grab
another hash_lock during buf_buddy_free.